requirements.txt
aiocfscrape==1.0.0
aiohttp==3.9.3
aiohttp-proxy==0.1.2
better-proxy==1.1.5
colorama==0.4.6
loguru==0.7.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
Pyrogram==2.0.106
tgcrypto==1.2.5